refactor: move useGetUserBalance to data-provider and update imports
